control_rover/requirements.txt
RPi.GPIO
keyboard
pytest